refactor: add custom hook for fetching ProductCardDetail for each product
